+++ src/main/kotlin/app/revanced/patches/youtube/ad/video/patch/VideoAdsPatch.kt
@@ -13,37 +13,41 @@ import app.revanced.patcher.patch.annotations.DependsOn
 import app.revanced.patcher.patch.annotations.Patch
 import app.revanced.patcher.util.smali.ExternalLabel
 import app.revanced.extensions.YouTubeCompatibility
-import app.revanced.patches.youtube.ad.video.fingerprints.LoadAdsFingerprint
+import app.revanced.patches.youtube.ad.video.fingerprints.LoadVideoAdsFingerprint
+import app.revanced.patches.youtube.ad.video.fingerprints.ShowVideoAdsFingerprint
 import app.revanced.patches.youtube.misc.integrations.patch.IntegrationsPatch
 
 @Patch
-@DependsOn([IntegrationsPatch::class])
+@DependsOn([IntegrationsPatch::class, SettingsPatch::class])
 @Name("video-ads")
 @Description("Removes ads in the video player.")
 @YouTubeCompatibility
 @Version("0.0.1")
 class VideoAdsPatch : BytecodePatch(
     listOf(
-        LoadAdsFingerprint
+        LoadVideoAdsFingerprint,
+        ShowVideoAdsFingerprint,
     )
 ) {
     override fun execute(context: BytecodeContext): PatchResult {
 
-        with(LoadAdsFingerprint.result!!) {
-            val insertIndex = scanResult.patternScanResult!!.startIndex
-            with(mutableMethod) {
-                addInstructions(
-                    insertIndex,
-                    """ 
-                            invoke-static { }, Lapp/revanced/integrations/patches/VideoAdsPatch;->shouldShowAds()Z
-                            move-result v4
-                            if-nez v4, :show_video_ads
-                            return-object v9
-                         """,
-                    listOf(ExternalLabel("show_video_ads", instruction(insertIndex)))
-                )
-            }
-        }
+        val lithoAdsFingerprintMethod = LoadVideoAdsFingerprint.result!!.mutableMethod
+
+        lithoAdsFingerprintMethod.addInstructions(
+            0, """
+                invoke-static { }, Lapp/revanced/integrations/patches/VideoAdsPatch;->shouldShowAds()Z
+                move-result v0
+                if-nez v0, :show_video_ads
+                return-void
+            """, listOf(ExternalLabel("show_video_ads", lithoAdsFingerprintMethod.instruction(0)))
+        )
+
+        ShowVideoAdsFingerprint.result!!.mutableMethod.addInstructions(
+            0, """
+                invoke-static { }, Lapp/revanced/integrations/patches/VideoAdsPatch;->shouldShowAds()Z
+                move-result v1
+            """
+        )
 
         return PatchResultSuccess()
     }